+++ src/js/components/PageHome/PageHome.jsx
@@ -16,8 +16,6 @@ import style from './PageHome.module.scss';
 // COMPONENT
 // ======================================================================
 
-const isLocal = process.env.REACT_APP_ENV === 'local';
-
 export const PageHome = () => {
   const dispatch = useDispatch();
   const downloadsRef = useRef(null);
@@ -55,8 +53,8 @@ export const PageHome = () => {
 
         <div className={style.body}>
           <p>
-            Chromatix is a desktop music player for Plex, that transforms your listening experience and makes
-            interacting with your music libraries a joy.
+            Chromatix is a desktop music player for Plex and Jellyfin that transforms your listening experience and
+            makes interacting with your music libraries a joy.
           </p>
         </div>
 
@@ -67,7 +65,7 @@ export const PageHome = () => {
             <Button onClick={scrollToDownloads} icon={<Icon icon="DownloadIcon" cover stroke strokeWidth={2} />}>
               Download the App
             </Button>
-            <div className={isLocal ? 'mt-20' : 'mt-20'}></div>
+            <div className="mt-20"></div>
           </>
         )}
 
@@ -81,19 +79,15 @@ export const PageHome = () => {
           >
             Login with Plex
           </Button>
-          {isLocal && (
-            <>
-              <Button
-                to="/login-jellyfin"
-                color={isElectron ? 'primary' : 'tertiary'}
-                size={isElectron ? 'large' : 'medium'}
-                wrap={false}
-                icon={<Icon icon="JellyfinSiteIcon" cover />}
-              >
-                Login with Jellyfin
-              </Button>
-            </>
-          )}
+          <Button
+            to="/login-jellyfin"
+            color={isElectron ? 'primary' : 'tertiary'}
+            size={isElectron ? 'large' : 'medium'}
+            wrap={false}
+            icon={<Icon icon="JellyfinSiteIcon" cover />}
+          >
+            Login with Jellyfin
+          </Button>
         </div>
       </div>
 
@@ -116,7 +110,7 @@ export const PageHome = () => {
         <div className="mt-40"></div>
 
         <div className={style.body}>
-          <p>Login with your Plex account to begin experiencing your music library in a whole new way.</p>
+          <p>Login with your Plex or Jellyfin account to begin experiencing your music library in a whole new way.</p>
         </div>
 
         <div className="mt-50"></div>
@@ -131,19 +125,15 @@ export const PageHome = () => {
           >
             Login with Plex
           </Button>
-          {isLocal && (
-            <>
-              <Button
-                to="/login-jellyfin"
-                color={isElectron ? 'primary' : 'tertiary'}
-                size={isElectron ? 'large' : 'medium'}
-                wrap={false}
-                icon={<Icon icon="JellyfinSiteIcon" cover />}
-              >
-                Login with Jellyfin
-              </Button>
-            </>
-          )}
+          <Button
+            to="/login-jellyfin"
+            color={isElectron ? 'primary' : 'tertiary'}
+            size={isElectron ? 'large' : 'medium'}
+            wrap={false}
+            icon={<Icon icon="JellyfinSiteIcon" cover />}
+          >
+            Login with Jellyfin
+          </Button>
         </div>
 
         {!isElectron && (