+++ src/js/services/jellyTools.js
@@ -22,6 +22,11 @@ const storageJellyUserKey = config.storageJellyUserKey;
 const storageServiceKey = config.storageServiceKey;
 const storageTokenKey = config.storageTokenKey;
 
+const trackFields =
+  'CustomRating,DateCreated,DateLastMediaAdded,DateLastRefreshed,DateLastSaved,Genres,MediaStreams,ProductionLocations,Tags,UserData';
+const artistAndAlbumFields =
+  'BackdropImageTags,CustomRating,DateCreated,DateLastMediaAdded,DateLastRefreshed,DateLastSaved,Genres,ProductionLocations,Tags,UserData';
+
 // ======================================================================
 // ENDPOINTS
 // ======================================================================
@@ -305,8 +310,7 @@ export const getAllArtists = ({ accessToken, genre, libraryId, serverBaseUrl, ta
             Tags: tag || null,
             SortBy: 'SortName',
             SortOrder: 'Ascending',
-            Fields:
-              'BackdropImageTags,CustomRating,DateCreated,DateLastMediaAdded,DateLastRefreshed,DateLastSaved,Genres,ProductionLocations,Tags,UserData',
+            Fields: artistAndAlbumFields,
             // StartIndex: 0,
             // Limit: 100
           },
@@ -357,8 +361,7 @@ export const getAllAlbumArtists = ({ accessToken, genre, libraryId, serverBaseUr
             Tags: tag || null,
             SortBy: 'SortName',
             SortOrder: 'Ascending',
-            Fields:
-              'BackdropImageTags,CustomRating,DateCreated,DateLastMediaAdded,DateLastRefreshed,DateLastSaved,Genres,ProductionLocations,Tags,UserData',
+            Fields: artistAndAlbumFields,
             // StartIndex: 0,
             // Limit: 100
           },
@@ -401,6 +404,9 @@ export const getArtistDetails = ({ accessToken, artistId, libraryId, serverBaseU
         .get(endpoint, {
           headers: getRequestHeaders(accessToken),
           signal: controller.signal,
+          params: {
+            Fields: artistAndAlbumFields,
+          },
         })
         .then((response) => {
           resolve(jellyTranspose.transposeArtistDetails(response, libraryId, serverBaseUrl, accessToken));
@@ -447,6 +453,7 @@ export const getAllArtistAlbums = ({ accessToken, artistId, libraryId, serverBas
             Recursive: true,
             SortBy: 'SortName',
             SortOrder: 'Ascending',
+            Fields: artistAndAlbumFields,
             // Filters: 'IsNotFolder', // Helps filter out compilation albums
             // ExcludeLocationTypes: 'Virtual', // Excludes virtual items, often compilations
           },
@@ -519,6 +526,7 @@ export const getAllArtistAppearanceAlbums = ({
             Recursive: true,
             SortBy: 'SortName',
             SortOrder: 'Ascending',
+            Fields: artistAndAlbumFields,
             // Filters: 'IsNotFolder', // Helps filter out compilation albums
             // ExcludeLocationTypes: 'Virtual', // Excludes virtual items, often compilations
           },
@@ -568,8 +576,7 @@ export const getAllArtistTracks = ({ accessToken, artistId, artistName, libraryI
             Recursive: true,
             SortBy: 'Album,SortName',
             SortOrder: 'Ascending',
-            Fields:
-              'CustomRating,DateCreated,DateLastMediaAdded,DateLastRefreshed,DateLastSaved,Genres,ProductionLocations,Tags,UserData',
+            Fields: trackFields,
             // Filters: 'IsNotFolder', // Helps filter out compilation albums
             // ExcludeLocationTypes: 'Virtual', // Excludes virtual items, often compilations
           },
@@ -620,8 +627,7 @@ export const getAllAlbums = ({ accessToken, genre, libraryId, serverBaseUrl, tag
             Tags: tag || null,
             SortBy: 'SortName',
             SortOrder: 'Ascending',
-            Fields:
-              'BackdropImageTags,CustomRating,DateCreated,DateLastMediaAdded,DateLastRefreshed,DateLastSaved,Genres,ProductionLocations,Tags,UserData',
+            Fields: artistAndAlbumFields,
             // StartIndex: 0,
             // Limit: 100
           },
@@ -664,6 +670,9 @@ export const getAlbumDetails = ({ accessToken, albumId, libraryId, serverBaseUrl
         .get(endpoint, {
           headers: getRequestHeaders(accessToken),
           signal: controller.signal,
+          params: {
+            Fields: artistAndAlbumFields,
+          },
         })
         .then((response) => {
           resolve(jellyTranspose.transposeAlbumDetails(response, libraryId, serverBaseUrl, accessToken));
@@ -708,8 +717,7 @@ export const getAlbumTracks = ({ accessToken, albumId, libraryId, serverBaseUrl,
             IncludeItemTypes: 'Audio',
             SortBy: 'ParentIndexNumber,IndexNumber,SortName',
             SortOrder: 'Ascending',
-            Fields:
-              'CustomRating,DateCreated,DateLastMediaAdded,DateLastRefreshed,DateLastSaved,Genres,ProductionLocations,Tags,UserData',
+            Fields: trackFields,
           },
         })
         .then((response) => {
@@ -769,8 +777,7 @@ export const getAllPlaylists = ({ accessToken, libraryId, serverBaseUrl, timeSta
             Recursive: true,
             SortBy: 'SortName',
             SortOrder: 'Ascending',
-            Fields:
-              'BackdropImageTags,CustomRating,DateCreated,DateLastMediaAdded,DateLastRefreshed,DateLastSaved,Genres,ProductionLocations,Tags,UserData',
+            Fields: artistAndAlbumFields,
             // StartIndex: 0,
             // Limit: 100
           },
@@ -813,6 +820,9 @@ export const getPlaylistDetails = ({ accessToken, libraryId, playlistId, serverB
         .get(endpoint, {
           headers: getRequestHeaders(accessToken),
           signal: controller.signal,
+          params: {
+            Fields: artistAndAlbumFields,
+          },
         })
         .then((response) => {
           resolve(jellyTranspose.transposePlaylistDetails(response, libraryId, serverBaseUrl, accessToken));
@@ -852,6 +862,9 @@ export const getPlaylistTracks = ({ accessToken, libraryId, playlistId, serverBa
         .get(endpoint, {
           headers: getRequestHeaders(accessToken),
           signal: controller.signal,
+          params: {
+            Fields: trackFields,
+          },
         })
         .then((response) => {
           resolve(jellyTranspose.transposeTrackArray(response, libraryId, serverBaseUrl, accessToken));