+++ src/js/utils/formatRecentDate.ts
@@ -7,6 +7,10 @@ import moment from 'moment';
  */
 
 const formatRecentDate = (timeStamp: number): string | null => {
+  if (!timeStamp) {
+    return null;
+  }
+
   // Convert timestamp to milliseconds and create a moment object
   const timeMoment = moment(timeStamp * 1000);